fixes1070: perf instrumentation — wall clock, and split the JS lump
Opening a perf session against the current instrument gives two answers
that cannot be acted on.

1. PERFACC's `total=` is the SUM OF THE INSTRUMENTED PHASES, not elapsed
   time. CLAUDE.md already records a PERFACC total=10s on a load timed at
   20-30s. So "js is 96% of the load" has really been "js is 96% of the
   part we happen to measure", and the unmeasured part may be larger.
   Added LIFE PERFWALL (wall / acct / unacct / unacctpct) off the existing
   nav-start t0. Read it first: a large unacct means no amount of tuning a
   measured phase moves the page load.

2. js=25.4s is one number. Bytecode caching (compile-bound) and a per-script
   budget (run-bound) are opposite pieces of work and it cannot tell them
   apart. js_exec now brackets compile and run separately via
   JS_EVAL_FLAG_COMPILE_ONLY + JS_EvalFunction. Not a behaviour change:
   __JS_EvalInternal ends in exactly that pair, and for JS_EVAL_TYPE_GLOBAL
   it has already set var_refs=NULL/sf=NULL — what JS_EvalFunction passes.

Also: LIFE JSTOP1..8, a per-script cost table (merge by name, evict
cheapest, sorted on emit); LIFE PERFN, per-phase CALL COUNTS taken free
inside the existing accum_* adders, so us/N separates "expensive once" from
"cheap 500 times"; and a GC timing hook in JS_RunGC.

All accumulate-into-statics, emitted once per navigation. Per-event logging
has polluted the measurement it was taken for here before (the fixes347d
var() trace was 95.5% of one load's log).

Two things the work turned up:

- Automatic cycle-GC has been OFF since fixes593, which pushed
  JS_SetGCThreshold past the memory cap to dodge a suspected collector
  double-free that was never root-caused. gcruns=0 is therefore the truth,
  not a broken hook — so JSPHASE carries gcarmed, and gc=0 can never read
  as "collection is free". It is also a live perf hypothesis: with the
  collector off, cyclic garbage survives the whole navigation.
- The harness stub for macos9_micros returned unsigned long; the real one
  returns double and every caller declares it that way. Integer in rax,
  caller reads xmm0 — every timing the harness has taken was garbage.
  Nothing asserted on them, so it stayed invisible. Same class as
  N_ELEMENTS and foreground_images.

Harness Test 49 asserts the split lands on the right side of the seam: one
compile-heavy script (4000 uncalled decls) and one run-heavy (2M-iteration
loop in ~40 bytes). Swapped brackets still produce two plausible non-zero
numbers and pass any "both > 0" check — this fails. 49 tests, ASan-clean,
leak count unchanged; check-macdefault OK.

+++ harness/harness_gen_stubs.c
@@ -12,7 +12,19 @@ void *macsurf_safe_realloc(void*p,size_t n){return realloc(p,n);}
 int macsurf_ptr_is_heap(const void*p){(void)p;return 1;}
 double macsurf_monotonic_ms(void){struct timespec t;clock_gettime(CLOCK_MONOTONIC,&t);return t.tv_sec*1000.0+t.tv_nsec/1e6;}
 unsigned long macsurf_get_ticks(void){struct timespec t;clock_gettime(CLOCK_MONOTONIC,&t);return (unsigned long)(t.tv_sec*60+t.tv_nsec/16666666);}
-unsigned long macos9_micros(void){struct timespec t;clock_gettime(CLOCK_MONOTONIC,&t);return (unsigned long)(t.tv_sec*1000000+t.tv_nsec/1000);}
+/* fixes1070 — this stub returned `unsigned long`. The REAL macos9_micros
+ * (main.c:122) returns DOUBLE, and every caller declares it that way with a
+ * local `extern double macos9_micros(void);` (macsurf_qjs.c, html.c,
+ * cssh_select.c, macos9_tls_fetcher.c). An integer return comes back in rax
+ * while the caller reads xmm0, so every timing this harness has ever taken
+ * was uninitialised-register garbage rather than a measurement.
+ *
+ * Nothing asserted on those numbers, so it stayed invisible — the exact
+ * harness false-signal class already documented for N_ELEMENTS and
+ * foreground_images. It matters now: the perf brackets added this round are
+ * meant to be developed and sanity-checked here before hardware sees them,
+ * and garbage in equals confident nonsense out. Match the real signature. */
+double macos9_micros(void){struct timespec t;clock_gettime(CLOCK_MONOTONIC,&t);return t.tv_sec*1000000.0+t.tv_nsec/1000.0;}
 /* guit + nsoptions[] are now REAL, properly-typed globals in driver.c
  * (driver.c needs a working guit->misc->schedule to pump box construction,
  * and box_get_style reads nsoptions[NSOPTION_author_level_css] at runtime —
